* Fixed some compiler complaints, comparing signed and unsigned values. * We can now add sound to spells in LUA (DAWN-88).   - setSoundSpellCasting( soundFile ) - is played at spell casting.   - setSoundSpellStart( soundFile ) - is played at spell start.   - setSoundSpellHit( soundFile ) - is played once the spell hits the target. * Cleaning up some whitespace / tab issues in the source files. * New spell for wolf and spider to use as melee (so they won't use the warrior melee sound). * Several spells has some new sound effect. Much more to add.. ;) * Fixed bug where the arinox dungeon would allow saving in the dungeon. * Possible to stop sound playback now in SoundEngine::stopSound( std::string soundFile).

+++ src/main.cpp
@@ -628,7 +628,7 @@ public:
 		}
 
         activity = ActivityType::Dying;
-        int numOfMoves = 0;
+        size_t numOfMoves = 0;
         if ( player->getClass() == CharacterClass::Warrior ) {
             numOfMoves = 9;
         } else if ( player->getClass() == CharacterClass::Ranger ) {
